查詢邏輯，第一個測試案例完成 四月份預算九千元 20170405-20170414，預算三千元

+++ GOOS_Sample/Services/IBudgetService.cs
@@ -17,7 +17,7 @@ namespace GOOS_Sample.Services
         event EventHandler Created;
         event EventHandler Updated;
 
-        int TotalBudget(Period period);
+        decimal TotalBudget(Period period);
     }
 
 }
